refactor: move `ConversationGuard::wipe()` out of `decrypt` module
Even though it's currently just used there, the `decrypt` module seems
like the wrong place for a general method like `wipe()`.
